Programmatically identifying optimal threshold sensitivity
The code now sees what the average object area is for each thresholding value, and uses this knowledge to select the value whose resulting object area is the closest to the desired (prior known) nuclear area. Makes the assumption that the fluorescent tags do not modify nuclear area in comparison to wt strains.